Improve water, add sand

+++ CloneCraft/Generator/OctavePerlin.h
@@ -8,7 +8,7 @@
 class OctavePerlin
 {
 public:
-	OctavePerlin(int octaves = 4, double persistence = 0.5, double frequency = 1.0);
+	OctavePerlin(int octaves = 4, float persistence = 0.5, float frequency = 1.0);
 	~OctavePerlin();
 
 	double getNoise(vec2 pos);
@@ -17,7 +17,7 @@ private:
 	std::vector<PerlinNoise> noises;
 
 	int m_octaves; // The more octaves, the more accurate is the noise
-	double m_persistence; // The less persistence, the less next octaves will have impact
-	double m_frequency; // The more frequency, the more dense is the noise
+	float m_persistence; // The less persistence, the less next octaves will have impact
+	float m_frequency; // The more frequency, the more dense is the noise
 };
